+++ src/app/alerts/index.ts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2022 The Pybricks Authors
+
+import { updateServerFailure } from './UpdateServerFailure';
+
+// gathers all of the alert creation functions for passing up to the top level
+export default {
+    updateServerFailure,
+};