+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+### Fixed
+- Fixed app freezing when checking for updates and update server is unreachable ([pybricks-code#1299]).
+
+[pybricks-code#1299]: https://github.com/pybricks/pybricks-code/issues/1299
+
 ## [2.0.0-beta.10] - 2022-11-11
 
 ### Added